refactor(chat): extract more operator surface helpers
