+++ crates/app/src/chat/operator_surfaces.rs
@@ -1,5 +1,3 @@
-use super::*;
-
 #[cfg(feature = "memory-sqlite")]
 use std::collections::BTreeSet;
 
@@ -8,11 +6,126 @@ use loongclaw_contracts::Capability;
 #[cfg(feature = "memory-sqlite")]
 use serde_json::json;
 
-use crate::conversation::{
-    collect_context_engine_runtime_snapshot, resolve_context_engine_selection,
-};
+use crate::CliResult;
+use crate::acp::resolve_acp_backend_selection;
+use crate::config;
+use crate::config::LoongClawConfig;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use crate::conversation::ContextCompactionReport;
+use crate::conversation::ConversationRuntimeBinding;
+use crate::conversation::ConversationTurnCoordinator;
+use crate::conversation::collect_context_engine_runtime_snapshot;
+use crate::conversation::resolve_context_engine_selection;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use crate::memory;
+#[cfg(feature = "memory-sqlite")]
+use crate::memory::runtime_config::MemoryRuntimeConfig;
+#[cfg(any(test, feature = "memory-sqlite"))]
 use crate::runtime_self_continuity;
+use crate::tui_surface::TuiActionSpec;
+use crate::tui_surface::TuiCalloutTone;
+use crate::tui_surface::TuiChoiceSpec;
+use crate::tui_surface::TuiHeaderStyle;
+use crate::tui_surface::TuiKeyValueSpec;
+use crate::tui_surface::TuiMessageSpec;
+use crate::tui_surface::TuiScreenSpec;
+use crate::tui_surface::TuiSectionSpec;
+use crate::tui_surface::render_tui_message_spec;
+use crate::tui_surface::render_tui_screen_spec;
 
+use super::CLI_CHAT_COMPACT_COMMAND;
+use super::CLI_CHAT_HELP_COMMAND;
+use super::CLI_CHAT_HISTORY_COMMAND;
+use super::CLI_CHAT_STATUS_COMMAND;
+use super::CLI_CHAT_TURN_CHECKPOINT_REPAIR_COMMAND;
+use super::CLI_CHAT_TURN_CHECKPOINT_REPAIR_COMMAND_ALIAS;
+use super::CliChatOptions;
+use super::CliTurnRuntime;
+use super::DEFAULT_FIRST_PROMPT;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use super::TurnCheckpointDiagnostics;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use super::TurnCheckpointDurabilityRenderLabels;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use super::TurnCheckpointRecoveryRenderLabels;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use super::TurnCheckpointSummaryRenderLabels;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use super::bool_yes_no_value;
+use super::detect_cli_chat_render_width;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use super::format_turn_checkpoint_failure_step;
+use super::print_rendered_cli_chat_lines;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use super::recovery_callout_tone;
+#[cfg(not(feature = "memory-sqlite"))]
+use super::render_cli_chat_feature_unavailable_lines_with_width;
+#[cfg(any(test, feature = "memory-sqlite"))]
+use super::render_turn_checkpoint_health_error_lines_with_width;
+use super::tui_plain_item;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(super) struct CliChatStartupSummary {
+    pub(super) config_path: String,
+    pub(super) memory_label: String,
+    pub(super) session_id: String,
+    pub(super) context_engine_id: String,
+    pub(super) context_engine_source: String,
+    pub(super) compaction_enabled: bool,
+    pub(super) compaction_min_messages: Option<usize>,
+    pub(super) compaction_trigger_estimated_tokens: Option<usize>,
+    pub(super) compaction_preserve_recent_turns: usize,
+    pub(super) compaction_fail_open: bool,
+    pub(super) acp_enabled: bool,
+    pub(super) dispatch_enabled: bool,
+    pub(super) conversation_routing: String,
+    pub(super) allowed_channels: Vec<String>,
+    pub(super) acp_backend_id: String,
+    pub(super) acp_backend_source: String,
+    pub(super) explicit_acp_request: bool,
+    pub(super) event_stream_enabled: bool,
+    pub(super) bootstrap_mcp_servers: Vec<String>,
+    pub(super) working_directory: Option<String>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(super) struct ManualCompactionResult {
+    pub(super) status: ManualCompactionStatus,
+    pub(super) before_turns: usize,
+    pub(super) after_turns: usize,
+    pub(super) estimated_tokens_before: Option<usize>,
+    pub(super) estimated_tokens_after: Option<usize>,
+    pub(super) summary_headline: Option<String>,
+    pub(super) detail: String,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(super) enum ManualCompactionStatus {
+    Applied,
+    NoChange,
+    FailedOpen,
+}
+
+#[cfg(feature = "memory-sqlite")]
+#[derive(Debug, Clone, PartialEq, Eq)]
+struct ManualCompactionWindowSnapshot {
+    turns: Vec<memory::WindowTurn>,
+    turn_count: Option<usize>,
+}
+
+#[allow(clippy::print_stdout)] // CLI output
+pub(super) fn print_cli_chat_startup(
+    runtime: &CliTurnRuntime,
+    options: &CliChatOptions,
+) -> CliResult<()> {
+    let summary = build_cli_chat_startup_summary(runtime, options)?;
+    for line in render_cli_chat_startup_lines(&summary) {
+        println!("{line}");
+    }
+    Ok(())
+}
+
+#[allow(clippy::print_stdout)] // CLI output
 pub(super) async fn print_turn_checkpoint_startup_health(runtime: &CliTurnRuntime) {
     #[cfg(feature = "memory-sqlite")]
     let render_width = detect_cli_chat_render_width();
@@ -103,7 +216,6 @@ pub(super) fn build_cli_chat_startup_summary(
     let context_engine_runtime = collect_context_engine_runtime_snapshot(&runtime.config)?;
     let compaction = context_engine_runtime.compaction;
     let acp_selection = resolve_acp_backend_selection(&runtime.config);
-
     Ok(CliChatStartupSummary {
         config_path: runtime.resolved_path.display().to_string(),
         memory_label: runtime.memory_label.clone(),
@@ -140,9 +252,106 @@ pub(super) fn build_cli_chat_startup_summary(
     })
 }
 
-pub(super) fn render_cli_chat_startup_lines(summary: &CliChatStartupSummary) -> Vec<String> {
+fn render_cli_chat_startup_lines(summary: &CliChatStartupSummary) -> Vec<String> {
     let render_width = detect_cli_chat_render_width();
     render_cli_chat_startup_lines_with_width(summary, render_width)
+}
+
+pub(super) fn should_run_missing_config_onboard(read: usize, input: &str) -> bool {
+    if read == 0 {
+        return false;
+    }
+
+    let normalized_input = input.trim().to_ascii_lowercase();
+
+    if normalized_input.is_empty() {
+        return true;
+    }
+
+    matches!(normalized_input.as_str(), "y" | "yes")
+}
+
+pub(super) fn render_cli_chat_missing_config_lines_with_width(
+    onboard_hint: &str,
+    width: usize,
+) -> Vec<String> {
+    let screen_spec = build_cli_chat_missing_config_screen_spec(onboard_hint);
+    render_tui_screen_spec(&screen_spec, width, false)
+}
+
+fn build_cli_chat_missing_config_screen_spec(onboard_hint: &str) -> TuiScreenSpec {
+    let intro_lines = vec![
+        format!("Welcome to {}!", config::PRODUCT_DISPLAY_NAME),
+        "No configuration found for interactive chat.".to_owned(),
+    ];
+    let sections = vec![TuiSectionSpec::ActionGroup {
+        title: Some("setup command".to_owned()),
+        inline_title_when_wide: true,
+        items: vec![TuiActionSpec {
+            label: "start setup".to_owned(),
+            command: onboard_hint.to_owned(),
+        }],
+    }];
+    let choices = vec![
+        TuiChoiceSpec {
+            key: "y".to_owned(),
+            label: "run setup wizard".to_owned(),
+            detail_lines: vec!["Create a config now and return to interactive chat.".to_owned()],
+            recommended: true,
+        },
+        TuiChoiceSpec {
+            key: "n".to_owned(),
+            label: "skip for now".to_owned(),
+            detail_lines: vec!["Exit chat now and keep the setup command for later.".to_owned()],
+            recommended: false,
+        },
+    ];
+    let footer_lines = vec!["Press Enter to accept y.".to_owned()];
+
+    TuiScreenSpec {
+        header_style: TuiHeaderStyle::Compact,
+        subtitle: Some("interactive chat".to_owned()),
+        title: Some("setup required".to_owned()),
+        progress_line: None,
+        intro_lines,
+        sections,
+        choices,
+        footer_lines,
+    }
+}
+
+pub(super) fn render_cli_chat_missing_config_decline_lines_with_width(
+    onboard_hint: &str,
+    width: usize,
+) -> Vec<String> {
+    let message_spec = build_cli_chat_missing_config_decline_message_spec(onboard_hint);
+    render_tui_message_spec(&message_spec, width)
+}
+
+fn build_cli_chat_missing_config_decline_message_spec(onboard_hint: &str) -> TuiMessageSpec {
+    let setup_hint = format!("You can run '{onboard_hint}' later to get started.");
+    let sections = vec![
+        TuiSectionSpec::Callout {
+            tone: TuiCalloutTone::Info,
+            title: Some("setup skipped".to_owned()),
+            lines: vec![setup_hint],
+        },
+        TuiSectionSpec::ActionGroup {
+            title: Some("start later".to_owned()),
+            inline_title_when_wide: true,
+            items: vec![TuiActionSpec {
+                label: "setup command".to_owned(),
+                command: onboard_hint.to_owned(),
+            }],
+        },
+    ];
+
+    TuiMessageSpec {
+        role: "chat".to_owned(),
+        caption: Some("setup required".to_owned()),
+        sections,
+        footer_lines: Vec::new(),
+    }
 }
 
 pub(super) fn render_cli_chat_startup_lines_with_width(
@@ -162,21 +371,21 @@ pub(super) fn render_cli_chat_status_lines_with_width(
 }
 
 fn build_cli_chat_startup_screen_spec(summary: &CliChatStartupSummary) -> TuiScreenSpec {
-    let mut sections = vec![
-        TuiSectionSpec::ActionGroup {
-            title: Some("start here".to_owned()),
-            inline_title_when_wide: true,
-            items: vec![TuiActionSpec {
-                label: "first prompt".to_owned(),
-                command: DEFAULT_FIRST_PROMPT.to_owned(),
-            }],
-        },
-        TuiSectionSpec::Narrative {
-            title: None,
-            lines: vec!["- type your request, or use /help for commands".to_owned()],
-        },
-    ];
+    let first_prompt_action = TuiActionSpec {
+        label: "first prompt".to_owned(),
+        command: DEFAULT_FIRST_PROMPT.to_owned(),
+    };
+    let start_here_section = TuiSectionSpec::ActionGroup {
+        title: Some("start here".to_owned()),
+        inline_title_when_wide: true,
+        items: vec![first_prompt_action],
+    };
+    let narrative_section = TuiSectionSpec::Narrative {
+        title: None,
+        lines: vec!["- type your request, or use /help for commands".to_owned()],
+    };
     let runtime_sections = build_cli_chat_runtime_sections(summary);
+    let mut sections = vec![start_here_section, narrative_section];
     sections.extend(runtime_sections);
 
     TuiScreenSpec {
@@ -225,7 +434,7 @@ fn build_cli_chat_runtime_sections(summary: &CliChatStartupSummary) -> Vec<TuiSe
         .compaction_trigger_estimated_tokens
         .map(|value| value.to_string())
         .unwrap_or_else(|| "-".to_owned());
-    let runtime_line = format!(
+    let runtime_value = format!(
         "ACP enabled={} dispatch_enabled={} routing={} backend={} ({}) allowed_channels={allowed_channels}",
         summary.acp_enabled,
         summary.dispatch_enabled,
@@ -233,65 +442,36 @@ fn build_cli_chat_runtime_sections(summary: &CliChatStartupSummary) -> Vec<TuiSe
         summary.acp_backend_id,
         summary.acp_backend_source,
     );
-    let session_items = vec![
-        TuiKeyValueSpec::Plain {
-            key: "session".to_owned(),
-            value: summary.session_id.clone(),
-        },
-        TuiKeyValueSpec::Plain {
-            key: "config".to_owned(),
-            value: summary.config_path.clone(),
-        },
-        TuiKeyValueSpec::Plain {
-            key: "memory".to_owned(),
-            value: summary.memory_label.clone(),
-        },
-    ];
     let context_engine_value = format!(
         "{} ({})",
         summary.context_engine_id, summary.context_engine_source
     );
-    let runtime_items = vec![
-        TuiKeyValueSpec::Plain {
-            key: "context engine".to_owned(),
-            value: context_engine_value,
-        },
-        TuiKeyValueSpec::Plain {
-            key: "acp".to_owned(),
-            value: runtime_line,
-        },
-    ];
     let session_section = TuiSectionSpec::KeyValues {
         title: Some("session details".to_owned()),
-        items: session_items,
+        items: vec![
+            tui_plain_item("session", summary.session_id.clone()),
+            tui_plain_item("config", summary.config_path.clone()),
+            tui_plain_item("memory", summary.memory_label.clone()),
+        ],
     };
     let runtime_section = TuiSectionSpec::KeyValues {
         title: Some("runtime details".to_owned()),
-        items: runtime_items,
+        items: vec![
+            tui_plain_item("context engine", context_engine_value),
+            tui_plain_item("acp", runtime_value),
+        ],
     };
     let continuity_section = TuiSectionSpec::KeyValues {
         title: Some("continuity maintenance".to_owned()),
         items: vec![
-            TuiKeyValueSpec::Plain {
-                key: "compaction".to_owned(),
-                value: summary.compaction_enabled.to_string(),
-            },
-            TuiKeyValueSpec::Plain {
-                key: "min messages".to_owned(),
-                value: compaction_min_messages,
-            },
-            TuiKeyValueSpec::Plain {
-                key: "trigger tokens".to_owned(),
-                value: compaction_trigger_estimated_tokens,
-            },
-            TuiKeyValueSpec::Plain {
-                key: "preserve recent".to_owned(),
-                value: summary.compaction_preserve_recent_turns.to_string(),
-            },
-            TuiKeyValueSpec::Plain {
-                key: "fail open".to_owned(),
-                value: summary.compaction_fail_open.to_string(),
-            },
+            tui_plain_item("compaction", summary.compaction_enabled.to_string()),
+            tui_plain_item("min messages", compaction_min_messages),
+            tui_plain_item("trigger tokens", compaction_trigger_estimated_tokens),
+            tui_plain_item(
+                "preserve recent",
+                summary.compaction_preserve_recent_turns.to_string(),
+            ),
+            tui_plain_item("fail open", summary.compaction_fail_open.to_string()),
         ],
     };
     let mut sections = vec![session_section, runtime_section, continuity_section];
@@ -306,18 +486,19 @@ fn build_cli_chat_runtime_sections(summary: &CliChatStartupSummary) -> Vec<TuiSe
         } else {
             summary.bootstrap_mcp_servers.join(",")
         };
-        let cwd_label = summary.working_directory.as_deref().unwrap_or("-");
+        let working_directory = summary.working_directory.as_deref().unwrap_or("-");
         let override_lines = vec![
             format!("explicit request: {}", summary.explicit_acp_request),
             format!("event stream: {}", summary.event_stream_enabled),
             format!("bootstrap MCP servers: {bootstrap_label}"),
-            format!("working directory: {cwd_label}"),
+            format!("working directory: {working_directory}"),
         ];
-        sections.push(TuiSectionSpec::Callout {
+        let override_callout = TuiSectionSpec::Callout {
             tone: TuiCalloutTone::Info,
             title: Some("acp overrides".to_owned()),
             lines: override_lines,
-        });
+        };
+        sections.push(override_callout);
     }
 
     sections
@@ -367,38 +548,59 @@ fn build_cli_chat_help_message_spec() -> TuiMessageSpec {
             value: "quit chat".to_owned(),
         },
     ];
-    let note_lines = vec![
-        "Type any non-command text to send a normal assistant turn.".to_owned(),
-        "Use /status to inspect runtime maintenance settings without sending a turn.".to_owned(),
-        "Use /history to inspect the active memory window when a reply feels off.".to_owned(),
-        "Use /compact to checkpoint the active session before the next turn.".to_owned(),
-    ];
+    let usage_section = TuiSectionSpec::Callout {
+        tone: TuiCalloutTone::Info,
+        title: Some("usage notes".to_owned()),
+        lines: vec![
+            "Type any non-command text to send a normal assistant turn.".to_owned(),
+            "Use /status to inspect runtime maintenance settings without sending a turn."
+                .to_owned(),
+            "Use /history to inspect the active memory window when a reply feels off.".to_owned(),
+            "Use /compact to checkpoint the active session before the next turn.".to_owned(),
+        ],
+    };
+    let command_section = TuiSectionSpec::KeyValues {
+        title: Some("slash commands".to_owned()),
+        items: command_items,
+    };
 
     TuiMessageSpec {
         role: "chat".to_owned(),
         caption: Some("commands".to_owned()),
-        sections: vec![
-            TuiSectionSpec::KeyValues {
-                title: Some("slash commands".to_owned()),
-                items: command_items,
-            },
-            TuiSectionSpec::Callout {
-                tone: TuiCalloutTone::Info,
-                title: Some("usage notes".to_owned()),
-                lines: note_lines,
-            },
-        ],
+        sections: vec![command_section, usage_section],
         footer_lines: Vec::new(),
     }
 }
 
-pub(super) fn print_help() {
-    let render_width = detect_cli_chat_render_width();
-    let rendered_lines = render_cli_chat_help_lines_with_width(render_width);
-    print_rendered_cli_chat_lines(&rendered_lines);
+pub(super) fn render_cli_chat_history_lines_with_width(
+    session_id: &str,
+    limit: usize,
+    history_lines: &[String],
+    width: usize,
+) -> Vec<String> {
+    let message_spec = build_cli_chat_history_message_spec(session_id, limit, history_lines);
+    render_tui_message_spec(&message_spec, width)
 }
 
-#[cfg(any(test, feature = "memory-sqlite"))]
+fn build_cli_chat_history_message_spec(
+    session_id: &str,
+    limit: usize,
+    history_lines: &[String],
+) -> TuiMessageSpec {
+    let caption = format!("session={session_id} limit={limit}");
+    let history_section = TuiSectionSpec::Narrative {
+        title: Some("sliding window".to_owned()),
+        lines: history_lines.to_vec(),
+    };
+
+    TuiMessageSpec {
+        role: "history".to_owned(),
+        caption: Some(caption),
+        sections: vec![history_section],
+        footer_lines: Vec::new(),
+    }
+}
+
 pub(super) fn render_manual_compaction_lines_with_width(
     session_id: &str,
     result: &ManualCompactionResult,
@@ -408,7 +610,6 @@ pub(super) fn render_manual_compaction_lines_with_width(
     render_tui_message_spec(&message_spec, width)
 }
 
-#[cfg(any(test, feature = "memory-sqlite"))]
 fn build_manual_compaction_message_spec(
     session_id: &str,
     result: &ManualCompactionResult,
@@ -418,7 +619,6 @@ fn build_manual_compaction_message_spec(
     let estimated_tokens_before = format_manual_compaction_tokens(result.estimated_tokens_before);
     let estimated_tokens_after = format_manual_compaction_tokens(result.estimated_tokens_after);
     let tone = manual_compaction_tone(result.status);
-
     let result_section = TuiSectionSpec::KeyValues {
         title: Some("compaction result".to_owned()),
         items: vec![
@@ -436,7 +636,6 @@ fn build_manual_compaction_message_spec(
             ),
         ],
     };
-
     let detail_section = TuiSectionSpec::Callout {
         tone,
         title: Some("details".to_owned()),
@@ -451,7 +650,6 @@ fn build_manual_compaction_message_spec(
     }
 }
 
-#[cfg(any(test, feature = "memory-sqlite"))]
 fn format_manual_compaction_status(status: ManualCompactionStatus) -> &'static str {
     match status {
         ManualCompactionStatus::Applied => "applied",
@@ -460,16 +658,13 @@ fn format_manual_compaction_status(status: ManualCompactionStatus) -> &'static s
     }
 }
 
-#[cfg(any(test, feature = "memory-sqlite"))]
 fn format_manual_compaction_tokens(value: Option<usize>) -> String {
     let Some(value) = value else {
         return "-".to_owned();
     };
-
     value.to_string()
 }
 
-#[cfg(any(test, feature = "memory-sqlite"))]
 fn manual_compaction_tone(status: ManualCompactionStatus) -> TuiCalloutTone {
     match status {
         ManualCompactionStatus::Applied => TuiCalloutTone::Success,
@@ -478,6 +673,14 @@ fn manual_compaction_tone(status: ManualCompactionStatus) -> TuiCalloutTone {
     }
 }
 
+#[allow(clippy::print_stdout)] // CLI output
+pub(super) fn print_help() {
+    let render_width = detect_cli_chat_render_width();
+    let rendered_lines = render_cli_chat_help_lines_with_width(render_width);
+    print_rendered_cli_chat_lines(&rendered_lines);
+}
+
+#[allow(clippy::print_stdout)] // CLI output
 pub(super) async fn print_manual_compaction(runtime: &CliTurnRuntime) -> CliResult<()> {
     #[cfg(feature = "memory-sqlite")]
     {
@@ -493,7 +696,6 @@ pub(super) async fn print_manual_compaction(runtime: &CliTurnRuntime) -> CliResu
         let render_width = detect_cli_chat_render_width();
         let rendered_lines =
             render_manual_compaction_lines_with_width(&runtime.session_id, &result, render_width);
-
         print_rendered_cli_chat_lines(&rendered_lines);
         Ok(())
     }
@@ -505,6 +707,41 @@ pub(super) async fn print_manual_compaction(runtime: &CliTurnRuntime) -> CliResu
         let rendered_lines = render_cli_chat_feature_unavailable_lines_with_width(
             "compact",
             "manual compaction unavailable: memory-sqlite feature disabled",
+            render_width,
+        );
+        print_rendered_cli_chat_lines(&rendered_lines);
+        Ok(())
+    }
+}
+
+#[allow(clippy::print_stdout)] // CLI output
+pub(super) async fn print_history(
+    session_id: &str,
+    limit: usize,
+    binding: ConversationRuntimeBinding<'_>,
+    #[cfg(feature = "memory-sqlite")] memory_config: &MemoryRuntimeConfig,
+) -> CliResult<()> {
+    #[cfg(feature = "memory-sqlite")]
+    {
+        let history_lines = load_history_lines(session_id, limit, binding, memory_config).await?;
+        let render_width = detect_cli_chat_render_width();
+        let rendered_lines = render_cli_chat_history_lines_with_width(
+            session_id,
+            limit,
+            &history_lines,
+            render_width,
+        );
+        print_rendered_cli_chat_lines(&rendered_lines);
+        Ok(())
+    }
+
+    #[cfg(not(feature = "memory-sqlite"))]
+    {
+        let _ = (session_id, limit, binding);
+        let render_width = detect_cli_chat_render_width();
+        let rendered_lines = render_cli_chat_feature_unavailable_lines_with_width(
+            "history",
+            "history unavailable: memory-sqlite feature disabled",
             render_width,
         );
 
@@ -654,6 +891,151 @@ fn build_manual_compaction_detail(
     }
 }
 
+#[cfg(any(test, feature = "memory-sqlite"))]
+fn format_window_history_lines(turns: &[memory::WindowTurn]) -> Vec<String> {
+    if turns.is_empty() {
+        return vec!["(no history yet)".to_owned()];
+    }
+
+    turns
+        .iter()
+        .map(|turn| {
+            format!(
+                "[{}] {}: {}",
+                turn.ts.unwrap_or_default(),
+                turn.role,
+                turn.content
+            )
+        })
+        .collect()
+}
+
+#[cfg(any(test, feature = "memory-sqlite"))]
+fn format_prompt_context_history_lines(entries: &[memory::MemoryContextEntry]) -> Vec<String> {
+    if entries.is_empty() {
+        return vec!["(no history yet)".to_owned()];
+    }
+
+    let mut lines = Vec::new();
+    for entry in entries {
+        match entry.kind {
+            memory::MemoryContextKind::Profile => {
+                lines.push("[profile]".to_owned());
+                lines.push(entry.content.clone());
+            }
+            memory::MemoryContextKind::Summary => {
+                lines.push("[summary]".to_owned());
+                lines.push(entry.content.clone());
+            }
+            memory::MemoryContextKind::RetrievedMemory => {
+                lines.push("[retrieved_memory]".to_owned());
+                lines.push(entry.content.clone());
+            }
+            memory::MemoryContextKind::Turn => {
+                lines.push(format!("{}: {}", entry.role, entry.content));
+            }
+        }
+    }
+    lines
+}
+
+#[cfg(feature = "memory-sqlite")]
+pub(super) async fn load_history_lines(
+    session_id: &str,
+    limit: usize,
+    binding: ConversationRuntimeBinding<'_>,
+    memory_config: &MemoryRuntimeConfig,
+) -> CliResult<Vec<String>> {
+    if let Some(ctx) = binding.kernel_context() {
+        let request = memory::build_window_request(session_id, limit);
+        let caps = BTreeSet::from([Capability::MemoryRead]);
+        let outcome = ctx
+            .kernel
+            .execute_memory_core(ctx.pack_id(), &ctx.token, &caps, None, request)
+            .await
+            .map_err(|error| format!("load history via kernel failed: {error}"))?;
+        if outcome.status != "ok" {
+            return Err(format!(
+                "load history via kernel returned non-ok status: {}",
+                outcome.status
+            ));
+        }
+        let turns = memory::decode_window_turns(&outcome.payload);
+        return Ok(format_window_history_lines(&turns));
+    }
+
+    let entries = memory::load_prompt_context(session_id, memory_config)
+        .map_err(|error| format!("load history failed: {error}"))?;
+    Ok(format_prompt_context_history_lines(&entries))
+}
+
+pub(super) fn parse_safe_lane_summary_limit(
+    input: &str,
+    default_window: usize,
+) -> CliResult<Option<usize>> {
+    parse_summary_limit(
+        input,
+        default_window,
+        &["/safe_lane_summary", "/safe-lane-summary"],
+    )
+}
+
+pub(super) fn parse_fast_lane_summary_limit(
+    input: &str,
+    default_window: usize,
+) -> CliResult<Option<usize>> {
+    parse_summary_limit(
+        input,
+        default_window,
+        &["/fast_lane_summary", "/fast-lane-summary"],
+    )
+}
+
+pub(super) fn parse_summary_limit(
+    input: &str,
+    default_window: usize,
+    aliases: &[&str],
+) -> CliResult<Option<usize>> {
+    let Some(primary_alias) = aliases.first().copied() else {
+        return Ok(None);
+    };
+
+    let mut tokens = input.split_whitespace();
+    let Some(command) = tokens.next() else {
+        return Ok(None);
+    };
+    if !aliases.contains(&command) {
+        return Ok(None);
+    }
+
+    let usage = format!("usage: {primary_alias} [limit]");
+    let default_limit = default_window.saturating_mul(4).max(64);
+    let limit = match tokens.next() {
+        Some(raw) => raw
+            .parse::<usize>()
+            .map_err(|error| format!("invalid {primary_alias} limit `{raw}`: {error}; {usage}"))?,
+        None => default_limit,
+    };
+    if limit == 0 {
+        return Err(format!("invalid {primary_alias} limit `0`; {usage}"));
+    }
+    if tokens.next().is_some() {
+        return Err(usage);
+    }
+    Ok(Some(limit))
+}
+
+pub(super) fn parse_turn_checkpoint_summary_limit(
+    input: &str,
+    default_window: usize,
+) -> CliResult<Option<usize>> {
+    parse_summary_limit(
+        input,
+        default_window,
+        &["/turn_checkpoint_summary", "/turn-checkpoint-summary"],
+    )
+}
+
 pub(super) fn parse_exact_chat_command(
     input: &str,
     aliases: &[&str],
@@ -688,6 +1070,15 @@ pub(super) fn is_manual_compaction_command(input: &str) -> CliResult<bool> {
 pub(super) fn is_cli_chat_status_command(input: &str) -> CliResult<bool> {
     let aliases = [CLI_CHAT_STATUS_COMMAND];
     let usage = "usage: /status";
+    parse_exact_chat_command(input, &aliases, usage)
+}
+
+pub(super) fn is_turn_checkpoint_repair_command(input: &str) -> CliResult<bool> {
+    let aliases = [
+        CLI_CHAT_TURN_CHECKPOINT_REPAIR_COMMAND,
+        CLI_CHAT_TURN_CHECKPOINT_REPAIR_COMMAND_ALIAS,
+    ];
+    let usage = "usage: /turn_checkpoint_repair";
     parse_exact_chat_command(input, &aliases, usage)
 }
